api/requirements.txt
fastapi[standard]
python-multipart
passlib[bcrypt]
pydantic
pyjwt
mysql-connector-python
email-validator